+++ c3smembership/presentation/views/dues_2016.py
@@ -260,7 +260,8 @@ def send_dues16_invoice_email(request, m_id=None):
             start_quarter)
         message = Message(
             subject=email_subject,
-            sender=request.registry.settings['c3smembership.mailaddr'],
+            sender=request.registry.settings[
+                'c3smembership.notification_sender'],
             recipients=[member.email],
             body=email_body,
         )
@@ -273,7 +274,8 @@ def send_dues16_invoice_email(request, m_id=None):
                 make_dues_invoice_investing_email(member)
         message = Message(
             subject=email_subject,
-            sender=request.registry.settings['c3smembership.mailaddr'],
+            sender=request.registry.settings[
+                'c3smembership.notification_sender'],
             recipients=[member.email],
             body=email_body,
         )
@@ -791,7 +793,8 @@ def dues16_reduction(request):
 
     message = Message(
         subject=email_subject,
-        sender=request.registry.settings['c3smembership.mailaddr'],
+        sender=request.registry.settings[
+            'c3smembership.notification_sender'],
         recipients=[member.email],
         body=email_body,
     )