+++ c3smembership/presentation/views/dues_2017.py
@@ -267,7 +267,8 @@ def send_dues17_invoice_email(request, m_id=None):
             start_quarter)
         message = Message(
             subject=email_subject,
-            sender=request.registry.settings['c3smembership.mailaddr'],
+            sender=request.registry.settings[
+                'c3smembership.notification_sender'],
             recipients=[member.email],
             body=email_body,
         )
@@ -280,7 +281,8 @@ def send_dues17_invoice_email(request, m_id=None):
                 make_dues_invoice_investing_email(member)
         message = Message(
             subject=email_subject,
-            sender=request.registry.settings['c3smembership.mailaddr'],
+            sender=request.registry.settings[
+                'c3smembership.notification_sender'],
             recipients=[member.email],
             body=email_body,
         )
@@ -801,7 +803,8 @@ def dues17_reduction(request):
 
     message = Message(
         subject=email_subject,
-        sender=request.registry.settings['c3smembership.mailaddr'],
+        sender=request.registry.settings[
+            'c3smembership.notification_sender'],
         recipients=[member.email],
         body=email_body,
     )