+++ c3smembership/presentation/views/dues_2018.py
@@ -285,7 +285,8 @@ def send_dues18_invoice_email(request, m_id=None):
             start_quarter)
         message = Message(
             subject=email_subject,
-            sender=request.registry.settings['c3smembership.mailaddr'],
+            sender=request.registry.settings[
+                'c3smembership.notification_sender'],
             recipients=[member.email],
             body=email_body,
         )
@@ -298,7 +299,8 @@ def send_dues18_invoice_email(request, m_id=None):
                 make_dues_invoice_investing_email(member)
         message = Message(
             subject=email_subject,
-            sender=request.registry.settings['c3smembership.mailaddr'],
+            sender=request.registry.settings[
+                'c3smembership.notification_sender'],
             recipients=[member.email],
             body=email_body,
         )
@@ -795,7 +797,8 @@ def dues18_reduction(request):
 
     message = Message(
         subject=email_subject,
-        sender=request.registry.settings['c3smembership.mailaddr'],
+        sender=request.registry.settings[
+            'c3smembership.notification_sender'],
         recipients=[member.email],
         body=email_body,
     )