+++ c3smembership/tests/test_join.py
@@ -70,7 +70,10 @@ class TestViews(unittest.TestCase):
         self.config.include('pyramid_mailer.testing')
         self.config.registry.settings[
             'c3smembership.url'] = 'https://yes.c3s.cc'
-        self.config.registry.settings['c3smembership.mailaddr'] = 'test@example.com'
+        self.config.registry.settings['c3smembership.notification_sender'] = \
+            'test@example.com'
+        self.config.registry.settings['c3smembership.status_receiver'] = \
+            'test@example.com'
         self.config.registry.settings['testing.mail_to_console'] = 'false'
         self.config.registry.get_mailer = get_mailer