+++ c3smembership/tests/test_mail_mail_confirmation.py
@@ -76,7 +76,7 @@ class TestMailMailConfirmationViews(unittest.TestCase):
         self.config = testing.setUp()
         self.config.include('pyramid_mailer.testing')
         self.config.registry.settings['c3smembership.url'] = 'http://foo.com'
-        self.config.registry.settings['c3smembership.mailaddr'] = \
+        self.config.registry.settings['c3smembership.notification_sender'] = \
             'test@example.com'
         self.config.registry.settings['testing.mail_to_console'] = 'false'
         self.config.registry.get_mailer = get_mailer