+++ c3smembership/tests/test_membership_acquisition.py
@@ -78,7 +78,7 @@ class TestReminderViews(unittest.TestCase):
         self.config.include('pyramid_mailer.testing')
         self.config.include('c3smembership.presentation.pagination')
         self.config.registry.settings['testing.mail_to_console'] = 'false'
-        self.config.registry.settings['c3smembership.mailaddr'] = \
+        self.config.registry.settings['c3smembership.notification_sender'] = \
             'test@example.com'
         self.config.add_route('dashboard', '/')
         def dashboard_content_size_provider(filtering):