+++ c3smembership/tests/test_membership_certificate.py
@@ -108,7 +108,7 @@ class TestMembershipCertificateViews(unittest.TestCase):
         DBSession.remove()
         self.session = _initTestingDB()
         self.config.registry.settings['testing.mail_to_console'] = 'no'
-        self.config.registry.settings['c3smembership.mailaddr'] = \
+        self.config.registry.settings['c3smembership.notification_sender'] = \
             'test@example.com'
         # set this to true to see mail bodies, but:
         # tests will fail: no mail in outbox