+++ c3smembership/tests/test_membership_dues_2016.py
@@ -183,7 +183,8 @@ class TestDues16Views(unittest.TestCase):
         self.config.include('pyramid_mailer.testing')
         self.config.registry.settings[
             'c3smembership.url'] = 'https://yes.c3s.cc'
-        self.config.registry.settings['c3smembership.mailaddr'] = 'c@example.com'
+        self.config.registry.settings['c3smembership.notification_sender'] = \
+            'c@example.com'
         self.config.registry.settings['testing.mail_to_console'] = 'false'
 
         DBSession.remove()